+++ changelog.md
@@ -8,6 +8,7 @@
 - Added option to delete synchronized files of single directory archive with one click
 - Added file type specific icons in directory listings
 - Added option for use the date of external files in add-dialog
+- Added option to really import files in media library (this disabled all external files functions for these files)
 - Added Taskfile as third way to build plugin release
 - Added check for PHP strict usage on every release with PHPStan